Updates to menuButton.js.
Two major types of changes (that should be in separate commits but
aren't).

1. Get rid of a bunch of the documentation. In principle, jsdoc style is
good, but here it's just overly cumbersome.

2. Beginnings of shift in how buttons manage control flow. Originally,
a button was passed in a callback that it invoked when it was done doing
whatever it was doing. This doesn't work well if buttons can be invoked
either by scanning, or by pressing directly. In the former case we want
a callback, in the latter nothing happens when the button is
finished. To implement this, don't pass in a callback. Instead, just
have the button emit an event when it is finished. If scanning is
happening, the scanner will listen for this event and will continue
scanning when it has been received.