Change the organization of the menus module.
Does not change functionality, but is a substantial change.
Mirrors what is done in menu-button. There is a global table of
menu constructors keyed by name. A single menu maker is invoked with a
name and a spec. This menu maker then dispatches on the name to get the
correct constructor from the table.