Refactor the code to handle speech.
Instead of having a module "speech" that exports functions, the module
"speaker" exports a constructor that creates a speaker object. This
object is passed to other objects in the program and its methods are
invoked to make speech happen in the program.

The reason for this change is that the speaker now has to keep track of
what the current language is. For this, it needs to have access to the
settings object.